GQ develop US2025080801 enhancement split main is train evaluate and predict
